pyproject.toml
[project]
name = "skills-hub"
version = "4.0.1"
description = "AI Agent Skills Hub Utilities"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0.1",
    "rich>=15.0.0",
    "streamlit>=1.56.0",
    "typer>=0.24.2",
]

[project.scripts]
hub = "scripts.installer:main"

[tool.setuptools]
packages = ["scripts"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.3.0",
]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-v --cov=scripts --cov-report=term-missing"
testpaths = [
    "tests",
]

[tool.ruff]
exclude = [
    "scaffolding-expert/examples",
    ".venv",
]
